+++ src/main/java/com/banking/service/interfaces/AgreementService.java
@@ -4,6 +4,7 @@ import com.banking.entity.Agreement;
 import com.banking.entity.entityEnumerations.AgreementStatus;
 
 
+import java.math.BigDecimal;
 import java.util.List;
 import java.util.UUID;
 
@@ -13,7 +14,7 @@ public interface AgreementService {
 
     Agreement createAgreement(Agreement agreement);
 
-    List<Agreement> findAgreementByInterestRate(Double interestRate);
+    List<Agreement> findAgreementByInterestRate(BigDecimal interestRate);
 
     Agreement editAgreement(UUID id, Agreement agreementFE);
 
@@ -21,7 +22,7 @@ public interface AgreementService {
 
     List<Agreement> deleteAgreementsByStatus(AgreementStatus status);
 
-    Agreement changeSumById(UUID id, Double newSum);
+    Agreement changeSumById(UUID id, BigDecimal newSum);
 
     List<Agreement> showDeleted();
 }
