+++ src/main/java/com/banking/service/interfaces/TransactionService.java
@@ -3,6 +3,7 @@ package com.banking.service.interfaces;
 import com.banking.entity.Transaction;
 import com.banking.entity.entityEnumerations.TransactionType;
 
+import java.math.BigDecimal;
 import java.util.List;
 import java.util.Optional;
 import java.util.UUID;
@@ -14,7 +15,7 @@ public interface TransactionService {
 
     List<Transaction> findTransactionsByType(TransactionType type);
 
-    List<Transaction> findTransactionsByAmount(Double amount);
+    List<Transaction> findTransactionsByAmount(BigDecimal amount);
 
     List<Transaction> findTransactionsByDescription(String description);
 }
